+++ CHANGELOG.rst
@@ -2,6 +2,11 @@
 Changelog for package maliput_integration
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+0.1.3 (2022-09-15)
+------------------
+* Adds Lane::ToSegmentPosition query. (`#123 <https://github.com/maliput/maliput_integration/issues/123>`_)
+* Contributors: Franco Cipollone
+
 0.1.2 (2022-08-16)
 ------------------
 * Matches with BoundingRegion moving to maliput. (`#122 <https://github.com/maliput/maliput_integration/issues/122>`_)